Convert ProfileListSection to JavaScript
