+++ constants/mockProfiles.js
@@ -1,0 +1,99 @@
+export const profileViewProfiles = [
+    {
+        id: '1',
+        image: require('../assets/images/users/user11.png'),
+        name: 'Samantha Smith',
+        profession: 'Software Engineer',
+        distance: '3.0km',
+    },
+    {
+        id: '2',
+        image: require('../assets/images/users/user12.png'),
+        name: 'Matilda Burch',
+        profession: 'UI/UX Designer',
+        distance: '4.5km',
+    },
+    {
+        id: '3',
+        image: require('../assets/images/users/user13.png'),
+        name: 'James Lamb',
+        profession: 'Junior Front-End Developer',
+        distance: '5.5km',
+    },
+    {
+        id: '4',
+        image: require('../assets/images/users/user14.png'),
+        name: 'Antonia Mcdaniel',
+        profession: 'Senior Back-End Developer',
+        distance: '3.0km',
+    },
+    {
+        id: '5',
+        image: require('../assets/images/users/user12.png'),
+        name: 'Matilda Burch',
+        profession: 'UI/UX Designer',
+        distance: '4.5km',
+    },
+];
+
+export const searchResultProfiles = [
+    {
+        id: '1',
+        image: require('../assets/images/users/user11.png'),
+        name: 'Samantha Smith',
+        profession: 'Software Engineer',
+        distance: '3.0km',
+    },
+    {
+        id: '2',
+        image: require('../assets/images/users/user12.png'),
+        name: 'Matilda Burch',
+        profession: 'UI/UX Designer',
+        distance: '3.0km',
+    },
+    {
+        id: '3',
+        image: require('../assets/images/users/user13.png'),
+        name: 'James Lamb',
+        profession: 'Junior Front-End Developer',
+        distance: '3.0km',
+    },
+    {
+        id: '4',
+        image: require('../assets/images/users/user14.png'),
+        name: 'Antonia Mcdaniel',
+        profession: 'Senior Back-End Developer',
+        distance: '3.0km',
+    },
+];
+
+export const similarProfiles = [
+    {
+        id: '1',
+        image: require('../assets/images/users/user3.png'),
+    },
+    {
+        id: '2',
+        image: require('../assets/images/users/user4.png'),
+    },
+    {
+        id: '3',
+        image: require('../assets/images/users/user6.png'),
+    },
+    {
+        id: '4',
+        image: require('../assets/images/users/user7.png'),
+    },
+    {
+        id: '5',
+        image: require('../assets/images/users/user3.png'),
+    },
+    {
+        id: '6',
+        image: require('../assets/images/users/user4.png'),
+    },
+    {
+        id: '7',
+        image: require('../assets/images/users/user6.png'),
+    },
+];
